Introduce Features.
Features can be initally disabled or disabled, and can be configured to be
enabled via Shuffleboard. Commands can be configured to be scheduled only
if all controlling features are enabled by using Features.whenAllEnabled().

This allows us to add functionality to the robot that might not be fully tested
by having the functionality disabled by default.