Store PersistedConfiguration record registry outside of /Preferences (#58)
This change moves the NetworkTables keys which store which PersistedConfiguration names map to which record classes outside of /Preferences.

As of the 2025 WPILib code, Preferences has a listener that updates all new topics to be persistent (for backwards compatibility of old dashboards; see https://github.com/wpilibsuite/allwpilib/commit/87fc49c66). Due to this, we cannot store data that should not be persistent under /Preferences without risking a race condition where the topic is later marked as persistent.

Other changes:
* Delete ".registeredTo" topics that were persisted under /Preferences