Made methods non final, so we can override them (#93)
We need them overridable for the arm code.
Mainly because we need to implement a way to move to an arm position
based on what position we are currently.
```
if (given_position in current_position.safe_positions_to_move_to) { 
  move(given_position)
} else { dont_move() }
```

This could theoretically be achieved with a helper method checking the
safety of a given position, but I would rather the setSetpoint() do this
for the assurance that someone doesn't call setSetpoint() directly,
skipping the safety check.